Revert "refactor(api): split YTMusicClient to satisfy lint"
This reverts commit 7741b47930630e10e71210384fdceadd82d98b3f.
